The previous design was compact; however, it only offered a parameterized facade to the types of a particular UML implementation API This means that OTI scripts could have dependencies to methods & fields of a particular UML implementation (e.g. MagicDraw, Papyrus)
The new design features a genuine OTI UML API: each OTI UML API object is a wrapper for an implementation-specific UML API object.
This means that it is necessary to use a caching mechanism to map to/from implementation-specific UML API objects and their corresponding OTI UML API objects.
However, the significant advantage is that scripts written against the OTI UML API are strongly independent of an implementation-specific UML API.


git-svn-id: https://www.omgwiki.org/repos/TIWG/Eclipse%20Scala%20Projects/org.omg.oti@509 b05f3425-8805-43f2-8a1d-5852a8693556